Tested FluidPOVButton
